Re-enable the integration broker for the System context.
