+++ server/src/main/java/io/spine/server/integration/IntegrationBroker.java
@@ -26,28 +26,191 @@
 
 package io.spine.server.integration;
 
+import com.google.common.collect.ImmutableSet;
+import io.grpc.stub.StreamObserver;
+import io.spine.annotation.Internal;
+import io.spine.core.BoundedContextName;
 import io.spine.core.Event;
+import io.spine.protobuf.AnyPacker;
+import io.spine.server.BoundedContext;
 import io.spine.server.ContextAware;
+import io.spine.server.ServerEnvironment;
 import io.spine.server.event.EventDispatcher;
+import io.spine.server.transport.ChannelId;
+import io.spine.server.transport.Publisher;
+import io.spine.server.transport.PublisherHub;
+import io.spine.server.transport.Subscriber;
+import io.spine.server.transport.SubscriberHub;
+import io.spine.server.transport.TransportFactory;
+import io.spine.server.type.EventClass;
 import io.spine.server.type.EventEnvelope;
+import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
+
+import static com.google.common.collect.ImmutableSet.toImmutableSet;
+import static io.spine.base.Identifier.newUuid;
+import static io.spine.base.Identifier.pack;
+import static io.spine.server.integration.Channels.toChannelId;
 
 /**
- * A part of some Bounded Context which defines the means of inter-Bounded Context communication.
+ * Dispatches {@linkplain ExternalMessage external messages} from and to a Bounded Context with
+ * which this broker is associated.
  *
- * <p>In particular, processes the {@code external} event subscriptions, including configuration
- * and message exchange.
+ * <p>In a multi-component environment messages may travel across components from one
+ * Bounded Context to another. Accordingly, an {@code IntegrationBroker} is available as a part
+ * of a Bounded Context.
  *
- * @see DefaultIntegrationBroker
+ * <p>An {@code IntegrationBroker} is always based upon {@linkplain TransportFactory transport}
+ * that delivers messages from and to it. For several Bounded Contexts to communicate, their brokers
+ * have to share the transport. Typically, that would be a single message queue.
+ *
+ * <p>{@code IntegrationBroker}s communicate with each other in order to keep the whole list of
+ * wanted external events and their potential sources (Bounded Contexts) up-to-date.
+ * They use two special messages for this:
+ * <ul>
+ *     <li>{@link BoundedContextOnline} is sent when a broker (and through it, its Bounded Context)
+ *     is connected to the transport
+ *     <li>{@link ExternalEventsWanted} is sent
+ *     <ul>
+ *         <li>in response to {@link BoundedContextOnline} sent by other brokers;
+ *         <li>when the collection of external events is updated inside of a Bounded Context.
+ *     </ul>
+ * </ul>
+ *
+ * <p><b>Receiving</b>
+ *
+ * <p>The messages from external components received by an {@code IntegrationBroker} via
+ * the transport are propagated into the Bounded Context via the domestic {@code EventBus}.
+ *
+ * <p><b>Publishing</b>
+ *
+ * <p>The messages requested by other parties are published from the domestic
+ * {@code EventBus} with the help of {@linkplain DomesticEventPublisher special dispatcher}.
+ *
+ * <p><b>Sample Usage</b>
+ *
+ * <p>Bounded Context "Projects" has a projection with an event handler that is subscribed to an
+ * external event as follows:
+ * <pre>
+ * {@code
+ * public class ProjectListView extends Projection<...>  {
+ *
+ *     {@literal @}Subscribe
+ *      public void on(@External UserDeleted event) {
+ *          // Remove the projects that belong to this user.
+ *          // ...
+ *      }
+ *  }
+ *  }
+ * </pre>
+ *
+ * <p>Upon a registration of the corresponding repository for this projection in the context,
+ * the broker associated with that context is informed that one more external event is needed.
+ * It sends out an updated {@code ExternalEventsWanted} saying that {@code UserDeleted}
+ * events are needed too.
+ *
+ * <p>Let's say the second Context is "Users". Its broker will receive
+ * the {@code ExternalEventsWanted} sent by "Projects". To handle it, it will create a bridge
+ * between "Users"'s Event Bus (which may eventually be transmitting a {@code UserDeleted} event)
+ * and the {@linkplain ServerEnvironment#transportFactory() transport}.
+ *
+ * <p>Once {@code UserDeleted} is emitted locally in "Users" context, it will be received
+ * by this bridge (as well as other local dispatchers) and published to the external transport by
+ * its {@code IntegrationBroker}.
+ *
+ * <p>The integration broker on the "Projects" side will receive the {@code UserDeleted}
+ * external message. The event will be dispatched to the external event handler of the projection.
+ *
+ * <p><b>Limitations</b>
+ *
+ * <p>An event type may be consumed by many Contexts but produced only by a single Context. This
+ * means that each given event type belongs only to one Context. The ownership of an event type may
+ * be changed between the versions of a system, but may not be changed while events are travelling
+ * between the Contexts.
  */
-public interface IntegrationBroker extends ContextAware, AutoCloseable {
+@Internal
+@SuppressWarnings("OverlyCoupledClass")
+public final class IntegrationBroker implements ContextAware, AutoCloseable {
+
+    private final SubscriberHub subscriberHub;
+    private final PublisherHub publisherHub;
+
+    private @MonotonicNonNull BoundedContextName contextName;
+    private @MonotonicNonNull BusAdapter bus;
+
+    private @MonotonicNonNull ObserveWantedEvents observeWantedEvents;
+    private @MonotonicNonNull BroadcastWantedEvents broadcast;
+
+    public IntegrationBroker() {
+        TransportFactory transportFactory = ServerEnvironment
+                .instance()
+                .transportFactory();
+        this.subscriberHub = new SubscriberHub(transportFactory);
+        this.publisherHub = new PublisherHub(transportFactory);
+    }
+
+    @Override
+    public void registerWith(BoundedContext context) {
+        checkNotRegistered();
+
+        this.contextName = context.name();
+        this.bus = new BusAdapter(this, context.eventBus());
+
+        runEventsExchange();
+        observeFellowContexts();
+        declareOnlineStatus();
+    }
+
+    private void runEventsExchange() {
+        ChannelId channel = Channels.eventsWanted();
+        broadcast = new BroadcastWantedEvents(contextName, publisherHub.get(channel));
+        observeWantedEvents = new ObserveWantedEvents(contextName, bus);
+        subscriberHub.get(channel)
+                     .addObserver(observeWantedEvents);
+    }
+
+    private void declareOnlineStatus() {
+        BoundedContextOnline notification =
+                BoundedContextOnline.newBuilder()
+                        .setContext(contextName)
+                        .vBuild();
+        ExternalMessage externalMessage = ExternalMessages.of(notification);
+        publisherHub.get(Channels.statuses())
+                    .publish(pack(newUuid()), externalMessage);
+    }
+
+    private void observeFellowContexts() {
+        StreamObserver<ExternalMessage> observer =
+                new ObserveFellowBoundedContexts(contextName, broadcast);
+        subscriberHub.get(Channels.statuses())
+                     .addObserver(observer);
+    }
+
+    @Override
+    public boolean isRegistered() {
+        return contextName != null;
+    }
 
     /**
      * Publishes the given event for other Bounded Contexts.
      *
+     * <p>If the given event belongs to another context, does nothing. More formally, if there is
+     * a subscriber channel in this broker for events of such a type, those events are
+     * NOT published from this Context.
+     *
      * @param event
-     *         the event to publish
+     *         event to publish
      */
-    void publish(EventEnvelope event);
+    @Internal
+    public void publish(EventEnvelope event) {
+        ChannelId channelId = toChannelId(event.messageClass());
+        boolean wantedByOthers = !subscriberHub.hasChannel(channelId);
+        if (wantedByOthers) {
+            Event outerObject = event.outerObject();
+            ExternalMessage msg = ExternalMessages.of(outerObject, contextName);
+            Publisher channel = publisherHub.get(channelId);
+            channel.publish(AnyPacker.pack(event.id()), msg);
+        }
+    }
 
     /**
      * Registers a local dispatcher which is subscribed to {@code external} messages.
@@ -55,7 +218,16 @@ public interface IntegrationBroker extends ContextAware, AutoCloseable {
      * @param dispatcher
      *         the dispatcher to register
      */
-    void register(EventDispatcher dispatcher);
+    public void register(EventDispatcher dispatcher) {
+        Iterable<EventClass> receivedTypes = dispatcher.externalEventClasses();
+        for (EventClass cls : receivedTypes) {
+            ChannelId channelId = toChannelId(cls);
+            Subscriber subscriber = subscriberHub.get(channelId);
+            IncomingEventObserver observer = observerFor(cls);
+            subscriber.addObserver(observer);
+            notifyTypesChanged();
+        }
+    }
 
     /**
      * Unregisters a local dispatcher which should no longer be subscribed
@@ -64,5 +236,51 @@ public interface IntegrationBroker extends ContextAware, AutoCloseable {
      * @param dispatcher
      *         the dispatcher to unregister
      */
-    void unregister(EventDispatcher dispatcher);
+    public void unregister(EventDispatcher dispatcher) {
+        Iterable<EventClass> externalEvents = dispatcher.externalEventClasses();
+        for (EventClass cls : externalEvents) {
+            ChannelId channelId = toChannelId(cls);
+            Subscriber subscriber = subscriberHub.get(channelId);
+            IncomingEventObserver observer = observerFor(cls);
+            subscriber.removeObserver(observer);
+        }
+        subscriberHub.closeStaleChannels();
+    }
+
+    private IncomingEventObserver observerFor(EventClass eventType) {
+        IncomingEventObserver observer = new IncomingEventObserver(contextName, eventType, bus);
+        return observer;
+    }
+
+    /**
+     * Notifies other Bounded Contexts that this Bounded Context now requests
+     * a different set of event types.
+     *
+     * <p>Sends out an instance of {@link ExternalEventsWanted} for that purpose.
+     */
+    private void notifyTypesChanged() {
+        ImmutableSet<ExternalEventType> eventTypes = subscriberHub
+                .ids()
+                .stream()
+                .map(Channels::typeOfTransmittedEvents)
+                .collect(toImmutableSet());
+        broadcast.onEventsChanged(eventTypes);
+    }
+
+    /**
+     * Removes all subscriptions and closes all the underlying transport channels.
+     */
+    @Override
+    public void close() throws Exception {
+        observeWantedEvents.close();
+        notifyTypesChanged();
+
+        subscriberHub.close();
+        publisherHub.close();
+    }
+
+    @Override
+    public String toString() {
+        return "IntegrationBroker of <" + contextName.getValue() + '>';
+    }
 }
