--- server/src/main/java/io/spine/server/integration/NoOpBroker.java
@@ -1,84 +1,0 @@
-/*
- * Copyright 2021, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.integration;
-
-import io.spine.server.BoundedContext;
-import io.spine.server.event.EventDispatcher;
-import io.spine.server.type.EventEnvelope;
-
-/**
- * An {@code IntegrationBroker} that does nothing.
- *
- * <p>This instance should be used in case the integration broker as-a-feature
- * {@linkplain io.spine.server.ContextSpec#hasIntegrationBroker() is disabled}.
- */
-public enum NoOpBroker implements IntegrationBroker {
-    INSTANCE;
-
-    /**
-     * Does nothing.
-     */
-    @Override
-    public void registerWith(BoundedContext context) {
-    }
-
-    /**
-     * Always returns {@code false}.
-     */
-    @Override
-    public boolean isRegistered() {
-        return false;
-    }
-
-    /**
-     * Does nothing.
-     */
-    @Override
-    public void publish(EventEnvelope event) {
-    }
-
-    /**
-     * Does nothing.
-     */
-    @Override
-    public void register(EventDispatcher dispatcher) {
-    }
-
-    /**
-     * Does nothing.
-     */
-    @Override
-    public void unregister(EventDispatcher dispatcher) {
-    }
-
-    /**
-     * Does nothing.
-     */
-    @Override
-    public void close() throws Exception {
-    }
-}
